Renamed `app` files in each language to be consistent with idiom.  e.g. program.cs for C#, index.js for Javascript, and app.python for Python.  Per review feedback from @amulyavarote
