+++ bindings/go/http/README.md
@@ -8,7 +8,7 @@ Visit [this](https://docs.dapr.io/developing-applications/building-blocks/bindin
 
 This quickstart includes one service:
  
-- Go service `batch`
+- Go service `app`
 
 ### Run and initialize PostgreSQL container
 
@@ -34,7 +34,7 @@ name: Install Go dependencies
 -->
 
 ```bash
-go build batch.go
+go build app.go
 ```
 
 <!-- END_STEP -->
@@ -54,7 +54,7 @@ sleep: 15
 -->
     
 ```bash
-dapr run --app-id go-input-binding-http --app-port 6003 --dapr-http-port 3503 --dapr-grpc-port 60003 --components-path ../../components -- go run batch.go
+dapr run --app-id go-input-binding-http --app-port 6003 --dapr-http-port 3503 --dapr-grpc-port 60003 --components-path ../../components -- go run app.go
 ```
 
 <!-- END_STEP -->
