+++ bindings/javascript/http/package.json
@@ -2,7 +2,7 @@
     "name": "bindings",
     "version": "1.0.0",
     "description": "",
-    "main": "batch.js",
+    "main": "index.js",
     "type": "module",
     "keywords": [],
     "author": "",
