+++ bindings/python/sdk/README.md
@@ -8,7 +8,7 @@ Visit [this](https://docs.dapr.io/developing-applications/building-blocks/bindin
 
 This quickstart includes one service:
  
-- Python service `batch`
+- Python service `app`
 
 ### Run and initialize PostgreSQL container
 
@@ -55,7 +55,7 @@ sleep: 15
 -->
     
 ```bash
-dapr run --app-id python-binding-quickstart-sdk --app-port 50051 --components-path ../../components -- python3 batch.py
+dapr run --app-id python-binding-quickstart-sdk --app-port 50051 --components-path ../../components -- python3 app.py
 ```
 
 <!-- END_STEP -->
